pyproject.toml
[project]
name = "project-akhir-struktur-data-grup-b-raja-jambi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "streamlit>=1.52.1",
    "pandas>=2.2.0",
    "requests>=2.31.0",
    "openpyxl>=3.1.2",
    "google-generativeai>=0.4.0",
    "fpdf>=1.7.2",
    "matplotlib>=3.8.0",
    "pillow>=10.2.0",
    "altair>=5.2.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
    "pypdf>=4.1.0",
    "reportlab>=4.4.6",
]
